refactor: introduce MapStruct and restructure DTOs, entities and configuration

+++ src/main/java/com/aceleradev/nexorder/services/EmployeeService.java
@@ -1,8 +1,12 @@
 package com.aceleradev.nexorder.services;
 
-import com.aceleradev.nexorder.dtos.employee.EmployeeDto;
+import com.aceleradev.nexorder.dtos.employee.EmployeeCreateRequest;
+import com.aceleradev.nexorder.dtos.employee.EmployeeResponse;
+import com.aceleradev.nexorder.dtos.employee.EmployeeUpdateRequest;
 import com.aceleradev.nexorder.entities.Employee;
+import com.aceleradev.nexorder.mapper.EmployeeMapper;
 import com.aceleradev.nexorder.repositories.EmployeeRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,80 +19,45 @@ public class EmployeeService {
     @Autowired
     private EmployeeRepository repository;
 
-    public List<EmployeeDto> findAll() {
-        List<Employee> employees = repository.findAll();
-        EmployeeDto dto = new EmployeeDto();
-        return employees.stream().map(
-                it -> {
-                    dto.setName(it.getName());
-                    dto.setIsActive(it.getActive());
-                    dto.setPhoneNumber(it.getPhoneNumber());
-                    dto.setEmail(it.getEmail());
-                    dto.setDescription(it.getDescription());
-                    dto.setLastLogin(it.getLastLogin());
-                    dto.setBaseSalary(it.getBaseSalary());
-                    dto.setRole(it.getRole());
-                    return dto;
-                }
-        ).toList();
+    @Autowired
+    private EmployeeMapper employeeMapper;
+
+    public List<EmployeeResponse> findAll() {
+        return employeeMapper.toResponseList(repository.findAll());
     }
 
-    public EmployeeDto findById(Long id) {
-        Optional<Employee> employee = repository.findById(id);
-        EmployeeDto dto = new EmployeeDto();
-        Employee obj = employee.get();
+    public EmployeeResponse findById(Long id) {
+        Employee employee = repository.findById(id).
+                orElseThrow(() -> new EntityNotFoundException("Employee not found."));
 
-        dto.setName(obj.getName());
-        dto.setIsActive(obj.getActive());
-        dto.setPhoneNumber(obj.getPhoneNumber());
-        dto.setEmail(obj.getEmail());
-        dto.setDescription(obj.getDescription());
-        dto.setLastLogin(obj.getLastLogin());
-        dto.setBaseSalary(obj.getBaseSalary());
-        dto.setRole(obj.getRole());
-
-        return dto;
-
+        return employeeMapper.toResponse(employee);
     }
 
-    public void createEmployee(EmployeeDto employeeDto) {
-        Employee employee = new Employee();
-
-        employee.setName(employeeDto.getName());
-        employee.setPassword(employeeDto.getPassword());
-        employee.setActive(employeeDto.getIsActive());
-        employee.setPhoneNumber(employeeDto.getPhoneNumber());
-        employee.setEmail(employeeDto.getEmail());
-        employee.setDescription(employeeDto.getDescription());
-        employee.setLastLogin(employeeDto.getLastLogin());
-        employee.setBaseSalary(employeeDto.getBaseSalary());
-        employee.setRole(employeeDto.getRole());
-
+    public EmployeeResponse create(EmployeeCreateRequest request) {
+        Employee employee = employeeMapper.toEntity(request);
         repository.save(employee);
+
+        return employeeMapper.toResponse(employee);
     }
 
-    public void deleteEmployee(Long id) {
+    public void delete(Long id) {
+        if (!repository.existsById(id)){
+            throw new EntityNotFoundException("Employee not found.");
+        }
+
         repository.deleteById(id);
     }
 
-    public void updateEmployee(Long id, EmployeeDto employeeDto) {
-        Optional<Employee> employeeById = repository.findById(id);
+    public EmployeeResponse update(Long id, EmployeeUpdateRequest request) {
 
-        if (employeeById.isPresent()) {
-            Employee employee = employeeById.get();
-            employee.setName(employeeDto.getName());
-            employee.setPassword(employeeDto.getPassword());
-            employee.setActive(employeeDto.getIsActive());
-            employee.setPhoneNumber(employeeDto.getPhoneNumber());
-            employee.setEmail(employeeDto.getEmail());
-            employee.setDescription(employeeDto.getDescription());
-            employee.setLastLogin(employeeDto.getLastLogin());
-            employee.setBaseSalary(employeeDto.getBaseSalary());
-            employee.setRole(employeeDto.getRole());
+        Employee employee = repository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Employee not found."));
 
-            repository.save(employee);
-        } else {
-            throw new RuntimeException("Employee not found");
-        }
+        employeeMapper.updateEmployeeFromDto(request, employee);
+
+        repository.save(employee);
+
+        return employeeMapper.toResponse(employee);
+
     }
 }
